requirements.txt
Django>=5.0,<6.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0
pytest-django>=4.0
requests>=2.31.0
